threading: use a pthread conditional variable instead of semaphore
This allows more control over the exact behavior, removes the global nature
of the semaphore on some systems, and will allow the addition of a timed wait
to the Event class.  It appears to resolve the deadlocks reproduced in the
pool test on Mac OS X; hopefully on Linux as well.