nal: refactor nal code, marshall packets directly into output buffer
This removes two malloc/free for every NAL unit and removes yet another set of
memcopies at the end of the each frame encode. We're now writing the escaped
NAL packets directly into the buffer handed back to the user.

We preserve the max size of this output buffer to prevent having to do any
reallocations once the encoder is running.