level: add --high-tier and auto-configure VBV if --crf N --level M (refs #61)
This patch is a major overhaul of the level enforcement logic. The first obvious
difference is that the user may specify the tier. The second difference is that
x265 will no longer run any configuration that might generate non-compliant
bitstreams.

Any of these conditions will cause the encode to abort if a minimum decoder
level was specified:

* picture size is too large
* frame rate is too high
* constant QP rate control is configured
* the specified level does not exist

Further, if CRF was specified, we now configure VBV using the maximum CPB size
and bitrate for their level/tier (and issue a warning that the output may now
be non-deterministic).

Note that the encoder will still encode the minimum decoder level which covers
the encode parameters.  So even if you specify --level 5.1, we may signal the
stream as level 4.0-High if the stream should be decodable at that level.

This further fixes the CLI to allow --level-idc or the shortened --level, just
as it now also supports --high-tier or the shortened --high.