scaling: use anonymous enums for class constants
these didn't need to be integers, they didn't need storage.  They're
only used to declare arrays and enforce loop bounds.  This fixes a link
error in debug compiles on Linux.  While I was at it I improved the
comments and moved one definition to entropy.cpp since that was the only
place it was used.