sao: simplify entropy coder initialization [CHANGES OUTPUTS]
Remove the hack from the HM that preserves fract_bits from the first row to all
rows; it never made any sense and we don't pretend to be anywhere near close to
bit accurate with the HM anymore.

This commit removes the Entrppy coder in FrameFilter since it was only used by
SAO and moves it into the SAO class.  The startSaoEnc() method is removed
entirely and another call to resetEntropy() (an expensive method) is replaced
with a load of the frame encoder's m_initSliceContext