cu: move cost variables from TComDataCU to Mode
we use md.bestMode != NULL to determine if no best mode has been found yet, so
we do not have to initialize costs to MAX_INT64 to simulate that condition.
Hopefully this is more robust.

this changes outputs, in a very good way, so I think this fixed a hidden bug
or two