analysis: Throw error messages for invalid analysis data from previous encode
This patch does the following:
Save encode - In the analysis file, save the param values that are required
to match in the load encode for the data to be reused.

Load encode - Read the saved param values and compare them with the current
param values. Throw error if there are any mismatches in the param values.